+++ tests/display.rs
@@ -86,8 +86,10 @@ fn timestamp_query() {
 fn sync() {
     let display = support::build_display();
 
-    let fence = glium::SyncFence::new_if_supported(&display);
-    fence.map(|f| f.wait());
+    let fence = glium::SyncFence::new(&display);
+    if let Ok(fence) = fence {
+        fence.wait();
+    }
 
     display.assert_no_error(None);
 }