+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+ - Replaced `read` and `read_if_supported` with `read` that returns a `ReadError`.
+ - Added `buffer::is_buffer_read_supported`.
+ - Replaced `SyncFence::new_if_supported` and `new` with `new` that returns a `Result<SyncFence, SyncNotSupportedError>`.
+
 ## Version 0.7.1 (2015-07-14)
 
  - Glium now automatically calls `glDraw*BaseVertex` if it is supported.